requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
blinker==1.8.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.4
django-debug-toolbar==4.3.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
openapi-codec==1.3.2
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
stripe==9.4.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
yarl==1.9.4
